tests,utils: Move GetVMPodName to libpod
